backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
PyJWT==2.8.0
bcrypt==4.1.3
python-decouple==3.8
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.0
django-filter==24.2
